Use 32 bit counters for send/recieve

+++ kernel-async-ga/client.py
@@ -164,7 +164,7 @@ runner.memcpy_d2h(
     order=MemcpyOrder.ROW_MAJOR,
     nonblock=False,
 )
-recvN = memcpy_view(out_tensors_u32, np.dtype(np.uint16))
+recvN = memcpy_view(out_tensors_u32, np.dtype(np.uint32))
 print(recvN)
 
 print("recv counter S ========================================================")
@@ -184,7 +184,7 @@ runner.memcpy_d2h(
     order=MemcpyOrder.ROW_MAJOR,
     nonblock=False,
 )
-recvS = memcpy_view(out_tensors_u32, np.dtype(np.uint16))
+recvS = memcpy_view(out_tensors_u32, np.dtype(np.uint32))
 print(recvS)
 
 print("recv counter E ========================================================")
@@ -204,7 +204,7 @@ runner.memcpy_d2h(
     order=MemcpyOrder.ROW_MAJOR,
     nonblock=False,
 )
-recvE = memcpy_view(out_tensors_u32, np.dtype(np.uint16))
+recvE = memcpy_view(out_tensors_u32, np.dtype(np.uint32))
 print(recvE)
 
 print("recv counter W ========================================================")
@@ -224,7 +224,7 @@ runner.memcpy_d2h(
     order=MemcpyOrder.ROW_MAJOR,
     nonblock=False,
 )
-recvW = memcpy_view(out_tensors_u32, np.dtype(np.uint16))
+recvW = memcpy_view(out_tensors_u32, np.dtype(np.uint32))
 print(recvW)
 
 print("recv counter sum =====================================================")
@@ -249,7 +249,7 @@ runner.memcpy_d2h(
     order=MemcpyOrder.ROW_MAJOR,
     nonblock=False,
 )
-sendN = memcpy_view(out_tensors_u32, np.dtype(np.uint16))
+sendN = memcpy_view(out_tensors_u32, np.dtype(np.uint32))
 print(sendN)
 
 print("send counter S ========================================================")
@@ -269,7 +269,7 @@ runner.memcpy_d2h(
     order=MemcpyOrder.ROW_MAJOR,
     nonblock=False,
 )
-sendS = memcpy_view(out_tensors_u32, np.dtype(np.uint16))
+sendS = memcpy_view(out_tensors_u32, np.dtype(np.uint32))
 print(sendS)
 
 print("send counter E ========================================================")
@@ -289,7 +289,7 @@ runner.memcpy_d2h(
     order=MemcpyOrder.ROW_MAJOR,
     nonblock=False,
 )
-sendE = memcpy_view(out_tensors_u32, np.dtype(np.uint16))
+sendE = memcpy_view(out_tensors_u32, np.dtype(np.uint32))
 print(sendE)
 
 print("send counter W ========================================================")
@@ -309,7 +309,7 @@ runner.memcpy_d2h(
     order=MemcpyOrder.ROW_MAJOR,
     nonblock=False,
 )
-sendW = memcpy_view(out_tensors_u32, np.dtype(np.uint16))
+sendW = memcpy_view(out_tensors_u32, np.dtype(np.uint32))
 print(sendW)
 
 print("send counter sum =====================================================")
